+++ Nos/Models/Author+CoreDataClass.swift
@@ -234,6 +234,41 @@ public class Author: NosManagedObject {
         return fetchRequest
     }
     
+    @nonobjc func followsRequest() -> NSFetchRequest<Author> {
+        let fetchRequest = NSFetchRequest<Author>(entityName: "Author")
+        fetchRequest.sortDescriptors = [NSSortDescriptor(keyPath: \Author.hexadecimalPublicKey, ascending: false)]
+        fetchRequest.predicate = NSPredicate(
+            format: "ANY followers = %@",
+            self
+        )
+        return fetchRequest
+    }    
+    
+    func reportsReferencingFetchRequest() -> NSFetchRequest<Event> {
+        guard let hexadecimalPublicKey else {
+            return Event.emptyRequest()
+        }
+        let fetchRequest = NSFetchRequest<Event>(entityName: "Event")
+        fetchRequest.sortDescriptors = [NSSortDescriptor(keyPath: \Event.createdAt, ascending: false)]
+        fetchRequest.predicate = NSPredicate(
+            format: "(kind = %i) AND ANY authorReferences.pubkey = %@", 
+            EventKind.report.rawValue,
+            hexadecimalPublicKey
+        )
+        return fetchRequest
+    }
+    
+    @nonobjc func authoredReportsRequest() -> NSFetchRequest<Event> {
+        let fetchRequest = NSFetchRequest<Event>(entityName: "Event")
+        fetchRequest.sortDescriptors = [NSSortDescriptor(keyPath: \Event.createdAt, ascending: false)]
+        fetchRequest.predicate = NSPredicate(
+            format: "kind = %i AND author = %@",
+            EventKind.report.rawValue,
+            self
+        )
+        return fetchRequest
+    }
+
     @nonobjc public class func emptyRequest() -> NSFetchRequest<Author> {
         let fetchRequest = NSFetchRequest<Author>(entityName: "Author")
         fetchRequest.sortDescriptors = [NSSortDescriptor(keyPath: \Author.hexadecimalPublicKey, ascending: true)]