+++ Nos/Views/NoteCard.swift
@@ -15,7 +15,7 @@ import Dependencies
 ///
 /// Use this view inside MessageButton to have nice borders.
 struct NoteCard: View {
-
+    
     @ObservedObject var note: Event
     
     var style = CardStyle.compact
@@ -24,24 +24,20 @@ struct NoteCard: View {
     @State private var userTappedShowOutOfNetwork = false
     @State private var replyCount = 0
     @State private var replyAvatarURLs = [URL]()
-
+    @State private var reportingAuthors = [Author]()
+    @State private var reports = [Event]()
+    @StateObject private var warningController = NoteWarningController()
+    
     @Environment(\.managedObjectContext) private var viewContext
     @EnvironmentObject private var router: Router
     @EnvironmentObject private var relayService: RelayService
     @EnvironmentObject private var currentUser: CurrentUser
     @Dependency(\.persistenceController) var persistenceController
-
+    
     private var showFullMessage: Bool
     private let showReplyCount: Bool
     private var hideOutOfNetwork: Bool
     private var replyAction: ((Event) -> Void)?
-    
-    private var showContents: Bool {
-        !hideOutOfNetwork ||
-        userTappedShowOutOfNetwork ||
-        currentUser.socialGraph.contains(note.author?.hexadecimalPublicKey) ||
-        Event.discoverTabUserIdToInfo.keys.contains(note.author?.hexadecimalPublicKey ?? "")
-    }
     
     private var attributedReplies: AttributedString? {
         if replyCount == 0 {
@@ -60,14 +56,16 @@ struct NoteCard: View {
             return nil
         }
     }
-    
+
     init(
         note: Event,
         style: CardStyle = .compact,
         showFullMessage: Bool = false,
         hideOutOfNetwork: Bool = true,
         showReplyCount: Bool = true,
-        replyAction: ((Event) -> Void)? = nil
+        replyAction: ((Event) -> Void)? = nil,
+        reported: Bool = false,
+        labeled: Bool = false
     ) {
         self.note = note
         self.style = style
@@ -81,67 +79,62 @@ struct NoteCard: View {
         VStack(alignment: .leading, spacing: 0) {
             switch style {
             case .compact:
-                HStack(alignment: .center, spacing: 0) {
-                    if showContents, let author = note.author {
-                        Button {
-                            router.currentPath.wrappedValue.append(author)
-                        } label: {
-                            NoteCardHeader(note: note, author: author)
+                VStack(spacing: 0) {
+                    HStack(alignment: .center, spacing: 0) {
+                        if !warningController.showWarning, let author = note.author {
+                            Button {
+                                router.currentPath.wrappedValue.append(author)
+                            } label: {
+                                NoteCardHeader(note: note, author: author)
+                            }
+                            NoteOptionsButton(note: note)
+                        } else {
+                            Spacer()
                         }
-                        NoteOptionsButton(note: note)
-                    } else {
-                        Spacer()
                     }
-                }
-                Divider().overlay(Color.cardDivider).shadow(color: .cardDividerShadow, radius: 0, x: 0, y: 1)
-                Group {
-                    if note.isStub {
-                        HStack {
-                            Spacer()
-                            ProgressView().foregroundColor(.primaryTxt)
-                            Spacer()
+                    Divider().overlay(Color.cardDivider).shadow(color: .cardDividerShadow, radius: 0, x: 0, y: 1)
+                    Group {
+                        if note.isStub {
+                            HStack {
+                                Spacer()
+                                ProgressView().foregroundColor(.primaryTxt)
+                                Spacer()
+                            }
+                            .padding(30)
+                        } else {
+                            CompactNoteView(
+                                note: note, 
+                                showFullMessage: showFullMessage, 
+                                loadLinks: !warningController.showWarning
+                            )
+                            .blur(radius: warningController.showWarning ? 6 : 0)
+                            .frame(maxWidth: .infinity)
                         }
-                        .padding(30)
-                    } else if showContents {
-                        CompactNoteView(note: note, showFullMessage: showFullMessage)
-                    } else {
-                        VStack {
-                            Localized.outsideNetwork.view
-                                .font(.body)
-                                .foregroundColor(.secondaryText)
-                                .padding(15)
-                            SecondaryActionButton(title: Localized.show) {
-                                withAnimation {
-                                    userTappedShowOutOfNetwork = true
+                        BeveledSeparator()
+                        HStack(spacing: 0) {
+                            if showReplyCount {
+                                StackedAvatarsView(avatarUrls: replyAvatarURLs, size: 20, border: 0)
+                                    .padding(.trailing, 8)
+                                if let replies = attributedReplies {
+                                    Text(replies)
+                                        .font(.subheadline)
+                                        .foregroundColor(Color.secondaryText)
                                 }
                             }
-                            .padding(.bottom, 15)
-                        }
-                        .frame(maxWidth: .infinity)
-                    }
-                    BeveledSeparator()
-                    HStack(spacing: 0) {
-                        if showReplyCount {
-                            StackedAvatarsView(avatarUrls: replyAvatarURLs, size: 20, border: 0)
-                                .padding(.trailing, 8)
-                            if let replies = attributedReplies {
-                                Text(replies)
-                                    .font(.subheadline)
-                                    .foregroundColor(Color.secondaryText)
+                            Spacer()
+                            RepostButton(note: note) {
+                                await repostNote()
                             }
+                            LikeButton(note: note)
+                            ReplyButton(note: note, replyAction: replyAction)
                         }
-                        Spacer()
-                        
-                        RepostButton(note: note) {
-                            await repostNote()
-                        }
-                        
-                        LikeButton(note: note)
-
-                        ReplyButton(note: note, replyAction: replyAction)
+                        .padding(.leading, 13)
                     }
-                    .padding(.leading, 13)
                 }
+                .blur(radius: warningController.showWarning ? 6 : 0)
+                .opacity(warningController.showWarning ? 0.3 : 1)
+                .frame(minHeight: warningController.showWarning ? 300 : nil)
+                .overlay(WarningView(controller: warningController))
             case .golden:
                 if let author = note.author {
                     GoldenPostView(author: author, note: note)
@@ -150,10 +143,16 @@ struct NoteCard: View {
                 }
             }
         }
-        .task(priority: .userInitiated) {
+        .task {
+            warningController.note = note
+            warningController.shouldHideOutOfNetwork = hideOutOfNetwork
+        }
+        .task {
             if note.isStub {
                 _ = await relayService.requestEvent(with: note.identifier)
             } 
+            self.reportingAuthors = note.reportingAuthors(followedBy: currentUser)
+            // print(self.reportingAuthors)
         }
         .onAppear {
             Task(priority: .userInitiated) {
@@ -184,7 +183,7 @@ struct NoteCard: View {
         .listRowInsets(EdgeInsets())
         .cornerRadius(cornerRadius)
     }
-    
+
     func repostNote() async {
         guard let keyPair = currentUser.keyPair else {
             return